backend/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-import-export==4.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.2.0
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
tzdata==2024.2
